add. prettier setting

+++ contracts/Vote.sol
@@ -2,41 +2,44 @@ pragma solidity ^0.5.6;
 
 contract Vote {
     // Mappings
-    mapping(uint => VoteInfo) public voteInfoes;
-    mapping(uint => CandidateInfo) public candidateInfoes;
-    
+    mapping(uint256 => VoteInfo) public voteInfoes;
+    mapping(uint256 => CandidateInfo) public candidateInfoes;
+
     struct VoteInfo {
-        uint idx;
+        uint256 idx;
         string name;
-        uint[] candidateIdxes;
+        uint256[] candidateIdxes;
     }
     struct CandidateInfo {
-        uint idx;
+        uint256 idx;
         string name;
-        uint voteCnt;
+        uint256 voteCnt;
     }
 
-    function getVote(uint _idx) public view returns (uint voteIdx, string memory voteName, uint[] memory candIdxes) {
-      voteIdx = _idx;
-      voteName = voteInfoes[_idx].name;
-      candIdxes = voteInfoes[_idx].candidateIdxes;
+    // prettier-ignore
+    function getVote(uint _idx) public view returns (uint256 voteIdx, string memory voteName, uint256[] memory candIdxes) {
+        voteIdx = _idx;
+        voteName = voteInfoes[_idx].name;
+        candIdxes = voteInfoes[_idx].candidateIdxes;
     }
-    
-    function getCandidate(uint _idx) public view returns (uint candIdx, string memory candName, uint voteCnt) {
-      candIdx = _idx;
-      candName = candidateInfoes[_idx].name;
-      voteCnt = candidateInfoes[_idx].voteCnt;
+
+    // prettier-ignore
+    function getCandidate(uint256 _idx) public view returns (uint256 candIdx, string memory candName, uint256 voteCnt) {
+        candIdx = _idx;
+        candName = candidateInfoes[_idx].name;
+        voteCnt = candidateInfoes[_idx].voteCnt;
     }
-    
-    function addCandidate(uint _idx, string memory _name) public {
+
+    function addCandidate(uint256 _idx, string memory _name) public {
         candidateInfoes[_idx] = CandidateInfo(_idx, _name, 0);
     }
-    
-    function addVote(uint _idx, string memory _name, uint[] memory candidates) public {
+
+    // prettier-ignore
+    function addVote(uint256 _idx, string memory _name, uint256[] memory candidates) public {
         voteInfoes[_idx] = VoteInfo(_idx, _name, candidates);
     }
-    
-    function vote(uint _candIdx) public {
+
+    function vote(uint256 _candIdx) public {
         candidateInfoes[_candIdx].voteCnt += 1;
     }
 }
